Move math function from util
